In100gram/requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.2
django-stdimage==6.0.1
djangorestframework==3.14.0
future==0.18.3
gunicorn==20.1.0
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
